Move domain/ contents to Paragraph component
