+++ components/Paragraph/Paragraph.tsx
@@ -1,6 +1,6 @@
 import React, { PureComponent, ReactNode } from 'react'
 
-import { TypografyReactChildren as TypografReactNode } from '~/domain/typografy/reactChildren'
+import { TypografyReactChildren as TypografReactNode } from './typografy/reactChildren'
 
 export class Paragraph extends PureComponent<{ children: ReactNode }> {
   render() {
