+++ components/Paragraph/typografy/reactChildren/index.ts
@@ -1,9 +1,9 @@
 import React, { ReactChild, ReactElement, ReactNode } from 'react'
 
-import { getType, TYPES } from '~/domain/react/childType'
-import { ReactTextNode as TextNode } from '~/domain/react/textNode'
-import { Typografy } from '~/domain/typografy'
-import { typografTextWithNeighbors } from '~/domain/typografy/textWithNeighbors'
+import { getType, TYPES } from '../../childType'
+import { ReactTextNode as TextNode } from '../../textNode'
+import { Typografy } from '..'
+import { typografTextWithNeighbors } from '../textWithNeighbors'
 
 export class TypografyReactChildren {
   public static processElement(node: ReactNode) {
