+++ jest.config.js
@@ -110,7 +110,7 @@ module.exports = {
   // rootDir: null,
 
   // A list of paths to directories that Jest should use to search for files in
-  roots: ['<rootDir>/domain', '<rootDir>/components', '<rootDir>/pages'],
+  roots: ['<rootDir>/components', '<rootDir>/pages'],
 
   // Allows you to use a custom runner instead of Jest's default test runner
   // runner: "jest-runner",
